io: move LimitReader/LimitWriter to separate file
